# swagger.json
{
    "swagger":"2.0",
    "info":{
        "version": "1.0.0",
        "title": "API Monitor",
        "description": "Flymonitor documentation",
        "licence":{
            "name":"MIT",
            "url":"https://flybyte.com.br"
        }
    },
    "paths":
...
        "/auth/verifytoken":{
            "get":
            {
                "tags":["Auth Calls"],
                "summary": "Verify Token",
                "description":"Route responsible for receiving user login and generating the access token",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                }
            ]
        },
        "/auth/verification":{
            "get":
            {
                "tags":["Auth Calls"],
                "summary": "API online",
                "description":"API verification is online",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                }
            ]
        },
        "/client/minhacam/:emailclient":{
            "get":
            {
                "tags":["Client Calls"],
                "summary": "My cameras",
                "description":"Route responsible for receiving user login and generating the access token",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true           
                }
            ]
        },
...
        "/client/findordercam/:email":{
            "get":
            {
                "tags":["Client Calls"],
                "summary": "Insert ordem cameras",
                "description":"Route responsible for entering cameras order",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                }
            ]
        },
        "/client/minhacam/:email":{
            "get":
            {
                "tags":["Client Calls"],
                "summary": "Find ordem cameras",
                "description":"Route responsible for sending cameras order",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true           
                }
            ]
        },
...
        "/master/register":{
            "post":
            {
                "tags":["Master Calls"],
                "summary": "Register master",
                "description":"Route responsible for insert new user master",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "razaosocial":{"type":"string"},
                        "CNPJouCPF":{"type":"integer"},
                        "email":{"type":"string"},
                        "nome:":{"type":"string"},
                        "sobreNome":{"type":"string"},
                        "endereco":{"type":"string"},
                        "cidade":{"type":"string"},
                        "estado":{"type":"string"},
                        "cep":{"type":"integer"},
                        "plano":{"type":"string"},
                        "password":{"type":"string"},
                        "usellevel":{"type":"string"}
                    }              
                }
            ]
        },
        "/master/listclients":{
            "get":
            {
                "tags":["Master Calls"],
                "summary": "List clients",
                "description":"Route responsible for list clients",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                }
            ]
        },
        "/master/listadmin":{
            "get":
            {
                "tags":["Master Calls"],
                "summary": "List admin",
                "description":"Route responsible for list admins",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                }
            ]
        },
        "/master/profileadmin/:email":{
            "get":
            {
                "tags":["Master Calls"],
                "summary": "List profile admin",
                "description":"Route responsible for listing the admin profile",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                }
            ]
        },
        "/master/editadminprofile":{
            "post":
            {
                "tags":["Master Calls"],
                "summary": "Edit admin profile",
                "description":"Route responsible for insert new user master",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "razaosocial":{"type":"string"},
                        "CNPJouCPF":{"type":"integer"},
                        "email":{"type":"string"},
                        "nome:":{"type":"string"},
                        "sobreNome":{"type":"string"},
                        "endereco":{"type":"string"},
                        "cidade":{"type":"string"},
                        "estado":{"type":"string"},
                        "cep":{"type":"integer"},
                        "plano":{"type":"string"},
                        "password":{"type":"string"},
                        "usellevel":{"type":"string"}
                    }              
                }
            ]
        }
    }
}